Fix grenade job API startup for uwsgi
During a grenade upgrade test run, we updated the grenade test to migrate
to using uwsgi for the API during the from-Xena upgrade, but missed that
the main grenade upgrade script was starting the API without uwsgi.

This patch corrects the grenade upgrade script to run the API process
properly.

Change-Id: Ia54de92ec44641e16c20281cfa634599073f39e3