Fix tox docs env to have the correct dependencies
Pip is pulling in unstrained dependencies during the "sibblings" package
install step. Normally this is fine as the dependencies have already been
installed and it's just installing the package under test. However, with
designate there are a lot of overlapping dependencies and pip will end up
reinstalling an unconstrained version. In this case it was Jinja2.

Change-Id: Ie49f99ba8e86913471ee6cd078f5afd28282c0f5