Fix BIND9 log collection
Currently the bind logs are not viewable in the zuul logs page, they prompt
for download. The wrong systemd unit logs are being exported and they were being double gziped. This patch corrects both of those issues such that the BIND9 logs will be properly collected in the zuul jobs.

Change-Id: Ied091e7acd0cb2c837f8ee7335b7881225da0126