Enable cache_ok on custom sqlalchemy UUID type
This patch adds the "cache_ok = True" to the custom sqlalchemy UUID type in designate.
It resolves the warning:
designate/sqlalchemy/base.py:243: SAWarning: TypeDecorator UUID() will not produce a cache key because the ``cache_ok`` attribute is not set to True.  This can have significant performance implications including some performance degradations in comparison to prior SQLAlchemy versions.  Set this attribute to True if this type object's state is safe to use in a cache key, or False to disable this warning. (Background on this error at: https://sqlalche.me/e/14/cprf)

Change-Id: Ic3aca3836b331db495109f6b6672e5c8af1d4028