Fix pecan lookup_controller DeprecationWarning
This patch updates Designate for the new funtion signature on pecan
lookup_controller to include the request information if available.

This corrects the following warning we were seeing in Designate:
DeprecationWarning: The function signature for pecan.routing.lookup_controller is changing in the next version of pecan.
Please update to: `lookup_controller(self, obj, remainder, request)`.

Change-Id: Icb206e59c83eec984599d48b3d4719c36eb83734