Fix sqlalchemy table_names DeprecationWarning
This patch updates Designate to use the new method of getting the table names from a sqlalchemy engine. It replaces the use of tables_names() with get_table_names from the sqlalchemy inspector.
The DeprecationWarning was:
SADeprecationWarning: The Engine.table_names() method is deprecated and will be removed in a future release.  Please refer to Inspector.get_table_names(). (deprecated since: 1.4)
  self.assertEqual(table_names, self.storage.engine.table_names())

Change-Id: I99d01a552d24123681e11f746c2e1195634efbfd