Fix designate-manage pool update MissingProjectID
This patch fixes a bug where adding an additional pool to designate using the "designate-manage pool update" command may fail with an exception:

designate.exceptions.MissingProjectID: A project ID must be specified when not using a project scoped token.

There was an extra check added as part of the scoped token work that a project ID must be provided when creating pools. This was incorrect as pools are still valid with a None project ID as they are a system resource and not tied to a specific project.

This patch removes that check, but retains the RBAC check for and "admin" token.

Closes-Bug: #1986733
Change-Id: I7345d7ef505420767209ba037e3c8930a282d03f
(cherry picked from commit a03c4657c2bf957fef035807ef7df732f7a1c994)