Fix mdns test_handler setup
If the unit tests were run isolated from the larger test suite, the mdns test_handler unit tests would fail with "AssertionError: 'TRANSPORT' must not be None" because the messaging and RPC fixtures were not setup for the test suite.
This patch adds the required test fixtures to the mdns test_handler setUp().

Change-Id: I7edf1e147637b013f35ba47e4d6b92d53d42fe84