Add clean_zonefile to BIND9 pool.yaml doc
This patch adds an example of the "clean_zonefile" option in the BIND9 targets in the pools.yaml file. The option has existed in the code, but was not documented in any way.
It also enables this feature in the bind9 devstack plugin for testing and to reduce the storage utilization during test runs.

Change-Id: I70b92b400cfeb059d1a1889ffa72621e0011d8c6