Workaround issue with dnspython >= 2.3.0
The dnspython module >= 2.3.0 now validates the opcodes used when building DNS messages. This breaks Designate because designate is using an "unassigned"[1] opcode 14 for command/control messages inside the backend agents framework.

This patch adds a workaround to override the dnspython opcode enum to include
opcode 14. This will give us time to either remove the agent framework via deprecation or to change the agent framework protocol to not rely on unassigned opcode values.

[1] https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-5

Partial-Bug: #2002950

Change-Id: If10443b2e361aa1b467fb64124ad3c82540bcddd