Fix sharing a zone with the zone owner
There was a bug that allowed users to create a zone share with the zone
owner. This would then cause issues deleting the zone share as the zone
owner owns the NS and SOA recordsets in the zone.
This patch raises a BadRequest if the user attempts to create a zone
share for the zone owner.

Closes-Bug: #2011585
Change-Id: I1b56c492436821f650d1ba669614d92595d2f476
(cherry picked from commit edcd2e09989852da94dd1333fcab0eab244d09d7)