Fix bug with designate-manage pool update losing existing record data
When designate-manage pool update is used we trigger an AFXR and
use that information to rebuild the zone. This is however only valid
for secondary zones, as the AXFR does not contain all the internal
information required for a primary zone and that causes designate to
lose important internal record data.

Closes-Bug: #2036750
Change-Id: I77a5d05cd8f55e975a16cb386221d53b63131886