Fix AXFR loop when updating secondary zone
When a secondary zone is created or updated we trigger an AXFR,
this AXFR then goes ahead and updates the zone. This then causes
the AXFR code to get triggered again as it thinks the masters
has changed.

We fix this by ensuring that when we update the zone from the AXFR
code that we always reset the masters field.

Partial-Bug: #1856442
Change-Id: Ifc2efb5a6d1091ff16373f3cbe3846f5e4371272
(cherry picked from commit ecaff9e6d348ce3b47a4b4fb71896951274191cb)