Fix unit tests depending on /etc/resolv.conf
Two unit tests loaded dns-python code that reads from /etc/resolv.conf on the
test host. If there are no nameservers listed in that file, the test would fail
even though it is not needed for the test.
This patch mocks out the calls dns-python makes that read the resolver records.

Closes-bug: #2064916
Change-Id: Id1949eab3ff9d02150527725c5d305e914942fa8