Set a default min_ttl for central
This fixes a bug when importing a zone with a TTL of 0 (bad practice, but
possible per the RFC[1]). A value of zero would cause Designate to use the
central process configuration setting of min_ttl, which had no default value.
This led to "None" being used, which caused the import to fail with a SQL
validation error.

This patch adds a default of zero to the min_ttl configuration setting.

Closes-bug: #2064972
Closes-bug: #1969299

[1] https://tools.ietf.org/html/rfc2181#section-8

Change-Id: I4d1314b657e14edfd3817612549df95650b159bb