Remove next pagination link for floating IP list
First of all the floating IP list API does not
support pagination as it doesn't pass any kind
of limit or marker down to the central API so
we simply get all the floating IPs returned.

But the floating IP API still accepts the limit
and marker query parameters, if limit is not given
we default to the default_limit_v2 config option
that has default value of 20.

If a project has exactly 20 floating IPs we hit
the bug documented in [1] where the returned list
is exactly the size of the limit used.

In the openstacksdk [2] this raises an exception
when the last item is the marker in the next url.

In the designateclient OSC plugin [3] this will
loop forever and generate HTTP requests until
stopped.

We resolve this by making the next link configurable
in the options for the object adapters but default
`next` to True to keep backward compatibility and
then set this to false for the floating IP list.

We add that since we cannot set the `links` option
to False as that would change the returned API
response body by removing links from the body
entirely.

[1] https://github.com/openstack/designate/blob/595894be2202bfd98805a0ed87281e69c59b5c1d/designate/objects/adapters/api_v2/base.py#L120
[2] https://github.com/openstack/openstacksdk/blob/79fbbbabff69e8e40e725ebbb8c605d2cc70ccae/openstack/resource.py#L2080
[3] https://github.com/openstack/python-designateclient/blob/490b0d6e7623859704746fdfd6d67e0d3a737ac8/designateclient/v2/base.py#L36

Closes-Bug: #2106112
Change-Id: I60ed37719fd0411159b8ec4f68ca7b001eb8a7be
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>
(cherry picked from commit e80666a1ddb6a43c2760505d0020d392cde87d21)