Add `designate-manage service clean` cmd
Add `designate-manage service clean` cmd to allow clean stale services.
Any services not provide heartbeat within double heartbeat_interval time
needs to be consider as dead services. And should be removed ASAP.

And in any case, even service is actually alive (maybe in a crazy rpc
timeout case), it can alway update and create it's service status back.
So we're fine to delete the service status in this case too.

Closes-Bug: #2110262

Change-Id: I211d0a60aa19ab4c0ae0fbb808e2d9080ccbbedd
Signed-off-by: ricolin <rlin@vexxhost.com>