Add pyproject.toml to support pip 23.1
pip 23.1 removed the "setup.py install" fallback for projects that do
not have pyproject.toml and now uses a pyproject.toml which is vendored
in pip. pip 25.3 will go a step further and enforce presence of this
file.

To address that, this change adds the minimal pyproject.toml to enable
pbr to be properly used to build editable wheels.

See [1] and [2] for more details on the removal of the fallback support.

[1] https://pip.pypa.io/en/stable/news/#v23-1
[2] https://github.com/pypa/pip/issues/836

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Change-Id: I40f9b914af6a714fe2e9b4180c4a9727bbafad48